refactor:refactoring of type definitions
