+++ src/components/features/populations/populations.type.ts
@@ -1,15 +1,3 @@
-export type ApiResponse = {
-  statusCode?: string;
-  message: string | null;
-  description?: string;
-  result?: Prefectures[];
-};
-
-export type Prefectures = {
-  prefCode: number;
-  prefName: string;
-};
-
 export type PopulationResponse = {
   statusCode?: string;
   message: string | null;
