+++ src/components/features/prefectures/prefectures.type.ts
@@ -1,0 +1,6 @@
+export type PrefectureResponse = {
+  statusCode?: string;
+  message: string | null;
+  description?: string;
+  result?: { prefCode: number; prefName: string }[];
+};
